Fixed unit tests linkage issue
